+++ src/pages/Submissions/index.js
@@ -1,14 +1,25 @@
-import React, {useEffect, useState} from 'react'
+import React, { useEffect, useState } from 'react'
 import { useParams } from 'react-router'
 import Event from '../Events/Event'
 import SubmissionsTable from './SubmissionsTable'
 import SubmissionsPending from './SubmissionsPending'
+import { getUserIsOwner } from '../../utils/getUserIsOwner '
+import { getEnvironment } from '../../utils/environment'
 
-const environment = 'http://localhost:3001'
+
+const environment = getEnvironment()
 
 const Submissions = () => {
   const { eventId } = useParams()
-  const [ lectures, setLectures ] = useState()
+  const [event, setEvent] = useState({})
+  const [lectures, setLectures] = useState()
+
+  useEffect(() => {
+    fetch(`${environment}/events/${eventId}`)
+      .then(res => res.json())
+      .then(data => setEvent(data))
+      .catch(err => console.error(err, 'Nenhum evento por aqui!'))
+  }, [eventId])
 
   const getLectures = () => {
     fetch(`${environment}/lectures`)
@@ -30,9 +41,17 @@ const Submissions = () => {
 
   return (
     <div>
-      <Event />
-      <SubmissionsTable lectures={lectures}  />
-      <SubmissionsPending lectures={lectures} handleUpdateLecture={handleUpdateLecture}/>
+      <Event event={event} />
+      {
+        getUserIsOwner(event.userId) ?
+          <>
+            <SubmissionsTable lectures={lectures} />
+            <SubmissionsPending lectures={lectures} handleUpdateLecture={handleUpdateLecture} />
+          </>
+          :
+          ''
+      }
+
     </div>
   )
 }
