+++ src/utils/getUserIsOwner .js
@@ -1,0 +1,4 @@
+export function getUserIsOwner(userId) {
+    const userIdLogged = localStorage.getItem('userId')
+    return String(userId) === String(userIdLogged)
+}
